+++ examples/js/psa.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const process = require('process');
+const { generateMyAwesomeTemplate } = require('./templates/tsClass.template');
 
 const completions = [];
 const notifications = [];
@@ -10,6 +11,34 @@ if (type === 'Info') {
   console.log(JSON.stringify({
     supported_languages: ["JavaScript"],
     goto_element_filter: ["JS:STRING_LITERAL"],
+    templates: [
+      {
+        type: "single_file",
+        name: "my_awesome_template",
+        title: "My Awesome Template",
+        path_regex: "^\/src\/[^\/]+\/$",
+        fields: [
+          {
+            name: "className",
+            title: "Class Name",
+            type: "Text",
+            options: []
+          },
+          {
+            name: "abstract",
+            title: "Is Abstract",
+            type: "Checkbox",
+            options: []
+          },
+          {
+            name: "comment",
+            title: "Comment",
+            type: "Select",
+            options: ["OptionA", "OptionB", "OptionC"]
+          }
+        ]
+      }
+    ]
   }));
 
   process.exit(0)
@@ -17,6 +46,16 @@ if (type === 'Info') {
 
 const contextString = fs.readFileSync(process.env.PSA_CONTEXT).toString();
 const context = JSON.parse(contextString);
+
+if (type === 'GenerateFileFromTemplate') {
+  console.log(JSON.stringify({
+    'file_name': context['formFields']['className'] + '.class.js',
+    'content': generateMyAwesomeTemplate(context['formFields']),
+  }));
+
+  process.exit(0)
+}
+
 const language = process.env.PSA_LANGUAGE;
 
 if (language === 'JavaScript') {