+++ examples/js/templates/tsClass.template.js
@@ -1,0 +1,15 @@
+module.exports = {
+  generateMyAwesomeTemplate: (formFields) => {
+    const abstract = formFields['abstract'] ? '/* abstract */ ' : '';
+
+    return `${abstract}class ${formFields['className']} {
+    /**
+    * @param ${formFields['comment']}
+    */
+    ${formFields['className']}() {
+    }
+}
+
+`;
+  }
+}